+++ internal/server/media.go
@@ -1,24 +1,24 @@
 // Copyright 2023 defsub
 //
-// This file is part of Takeout.
+// This file is part of TakeoutFM.
 //
-// Takeout is free software: you can redistribute it and/or modify it under the
+// TakeoutFM is free software: you can redistribute it and/or modify it under the
 // terms of the GNU Affero General Public License as published by the Free
 // Software Foundation, either version 3 of the License, or (at your option)
 // any later version.
 //
-// Takeout is distributed in the hope that it will be useful, but WITHOUT ANY
+// TakeoutFM is distributed in the hope that it will be useful, but WITHOUT ANY
 // WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 // FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for
 // more details.
 //
 // You should have received a copy of the GNU Affero General Public License
-// along with Takeout.  If not, see <https://www.gnu.org/licenses/>.
+// along with TakeoutFM.  If not, see <https://www.gnu.org/licenses/>.
 
 package server
 
 import (
-	"fmt"
+	"strings"
 	"github.com/takeoutfm/takeout/internal/auth"
 	"github.com/takeoutfm/takeout/internal/config"
 	"github.com/takeoutfm/takeout/internal/music"
@@ -61,7 +61,7 @@ func mediaConfigFor(root *config.Config, user *auth.User) (string, *config.Confi
 }
 
 func mediaConfig(root *config.Config, mediaName string) (*config.Config, error) {
-	path := fmt.Sprintf("%s/%s", root.Server.MediaDir, mediaName)
+	path := strings.Join([]string{root.Server.MediaDir, mediaName}, "/")
 	// load relative media configuration
 	userConfig, err := config.LoadConfig(path)
 	if err != nil {