+++ internal/video/video.go
@@ -1,19 +1,19 @@
 // Copyright 2023 defsub
 //
-// This file is part of Takeout.
+// This file is part of TakeoutFM.
 //
-// Takeout is free software: you can redistribute it and/or modify it under the
+// TakeoutFM is free software: you can redistribute it and/or modify it under the
 // terms of the GNU Affero General Public License as published by the Free
 // Software Foundation, either version 3 of the License, or (at your option)
 // any later version.
 //
-// Takeout is distributed in the hope that it will be useful, but WITHOUT ANY
+// TakeoutFM is distributed in the hope that it will be useful, but WITHOUT ANY
 // WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 // FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for
 // more details.
 //
 // You should have received a copy of the GNU Affero General Public License
-// along with Takeout.  If not, see <https://www.gnu.org/licenses/>.
+// along with TakeoutFM.  If not, see <https://www.gnu.org/licenses/>.
 
 // Package video provides support for all movie media.
 package video
@@ -131,7 +131,7 @@ func (v *Video) Search(q string, limit ...int) []Movie {
 
 func (v *Video) MovieURL(m Movie) *url.URL {
 	// FIXME assume first bucket!!!
-	return v.buckets[0].Presign(m.Key)
+	return v.buckets[0].ObjectURL(m.Key)
 }
 
 func (v *Video) MoviePoster(m Movie) string {