+++ spiff/playlist.go
@@ -1,23 +1,23 @@
 // Copyright 2023 defsub
 //
-// This file is part of Takeout.
+// This file is part of TakeoutFM.
 //
-// Takeout is free software: you can redistribute it and/or modify it under the
+// TakeoutFM is free software: you can redistribute it and/or modify it under the
 // terms of the GNU Affero General Public License as published by the Free
 // Software Foundation, either version 3 of the License, or (at your option)
 // any later version.
 //
-// Takeout is distributed in the hope that it will be useful, but WITHOUT ANY
+// TakeoutFM is distributed in the hope that it will be useful, but WITHOUT ANY
 // WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 // FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for
 // more details.
 //
 // You should have received a copy of the GNU Affero General Public License
-// along with Takeout.  If not, see <https://www.gnu.org/licenses/>.
+// along with TakeoutFM.  If not, see <https://www.gnu.org/licenses/>.
 
-// Package spiff provides the model for all playlists used within Takeout. A
+// Package spiff provides the model for all playlists used within TakeoutFM. A
 // spiff is a container for one or more media tracks with basic media metadata.
-// Takeout also uses json patch to manipulate spiffs for remote playlist
+// TakeoutFM also uses json patch to manipulate spiffs for remote playlist
 // management.
 package spiff